+++ packages/page-alert/src/PageAlert.tsx
@@ -8,7 +8,7 @@ import {
 	WarningFilledIcon,
 } from '@ag.ds-next/icon';
 
-import { PageAlertTitle } from './page-alert-title';
+import { PageAlertTitle } from './PageAlertTitle';
 
 export type PageAlertTone = 'success' | 'error' | 'warning' | 'info';
